+++ gen_api/__init__.py
@@ -1,5 +1,5 @@
 from ._version import __version__
-from .api import *
-from .esp_api import *
-from .cat_api import *
-from .deu_api import *
+from .basics import *
+from .esp_basics import *
+from .cat_basics import *
+from .deu_basics import *
