+++ gen_api/deu_basics.py
